finish to add convergence data to missing implicitSetup

+++ src/coreComponents/physicsSolvers/PhysicsSolverBase.cpp
@@ -1179,7 +1179,7 @@ void PhysicsSolverBase::setupDofs( DomainPartition const & GEOS_UNUSED_PARAM( do
   GEOS_ERROR( "PhysicsSolverBase::setupDofs called!. Should be overridden." );
 }
 
-void PhysicsSolverBase::doSmthEarlyStep( integer const & time_n, real64 const & dt, real64 const & cycleNumber )
+void PhysicsSolverBase::doSmthEarlyStep( real64 const & time_n, real64 const & dt, integer const & cycleNumber )
 {
   getConvergenceStats().updateSolverStep( time_n, dt, cycleNumber );
 }
